requirements.txt
streamlit>=1.28.0
easyocr>=1.7.0
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
qwen-agent>=0.0.5
openai>=1.3.0
python-dotenv>=1.0.0
plotly>=5.17.0
streamlit-webrtc>=0.47.0
av>=10.0.0